+++ authentication_lib/src/main/java/org/octri/authentication/server/controller/UserController.java
@@ -1,12 +1,9 @@
 package org.octri.authentication.server.controller;
 
-import static org.octri.authentication.server.security.entity.PasswordResetToken.LONG_EXPIRE_IN_MINUTES;
-
 import java.util.Date;
 import java.util.List;
 import java.util.Optional;
 import java.util.Set;
-import java.time.Duration;
 
 import org.apache.commons.lang3.StringUtils;
 import org.apache.commons.logging.Log;
@@ -216,9 +213,8 @@ public class UserController {
 				// The new user is LDAP if table-based auth is not enabled or if LDAP was indicated in the form
 				Boolean ldapUser = !getTableBasedEnabled() || user.getLdapUser();
 				Boolean noemailProfileIsNotActive = !profileUtils.isActive(ProfileUtils.AuthProfile.noemail);
-				Duration passwordResetTokenExpiration = getPasswordTokenValidFor();
-				PasswordResetToken token = passwordResetTokenService.generatePasswordResetToken(savedUser, passwordResetTokenExpiration);
 				if (!ldapUser) {
+					PasswordResetToken token = passwordResetTokenService.generatePasswordResetToken(savedUser);
 					if (noemailProfileIsNotActive) {
 						userService.sendPasswordResetTokenEmail(token, request, true, false);
 					}
@@ -291,14 +287,6 @@ public class UserController {
 	public String getUsernameStyle() {
 		UsernameStyle usernameStyle = authenticationProperties.getUsernameStyle();
 		return usernameStyle != null ? usernameStyle.toString() : null;
-	}
-
-	/**
-	*@return the value for the password token validity
-	*/
-	@ModelAttribute("passwordTokenValidFor")
-	public Duration getPasswordTokenValidFor() {
-		return authenticationProperties.getPasswordTokenValidFor(); 
 	}
 
 	/**